It works!
This commit makes it possible to copy a directory of files to multiple
devices. There is not yet support for stopping the copy to
allow the user to switch disks.

This is a huge commit due to the big rapid changes, but it mostly
consists of a switch in logging libs from log15 to logrus. I like being
api compatible with the default log lib, and logrus also provides for
better future expansion.