Add TestFileSyncAcrossDevicesNoSplit
When syncing to a second device, "file not found" errors were ocurring
because the directory tree to the file was not being generated before
writing the file. The directory tree is generated on successive devices
if the file is being split across devices. This test checks for syncing
to multiple devices if no files are being split.