Improve command argument handling for config file
A change was made to the way getConfigFile() interprets the path passed
to it. If the path contains an extension, then "config.yml" is not
appended to the name.

Support for the GDS_CONFIG_DIR environment variable was also added. The
default command arguments reference $GDS_CONFIG_DIR for the base
directory to the config file path, the default configuration base
directory is set to $HOME/.config/gds.