Improve sync context JSON exporting
The getContextFile() function was added for making sure a file exists so
that the sync context data could be encoded as json and saved to the
file. By default, the context file is saved in the GDS_CONFIG_DIR with
an RFC3339 time format appended to the file name. The context file path
can be overridden with the "--context-file" command argument.