Fixed nasty file split bug in NewCatalog()
When a file is split across devices in NewCatalog(), the file, and its
parent directories are created on the next device, this can lead to
unexpected growth in the total backup size. The bug appeared when the
remaining file size + the size of the parent directories exceeded the
size of the remaining devicie pool space.

NewCatalog() was modified to return an
CatalogNotEnoughDevicePoolSpaceError type. This type has embedded
TotalCatalogSize and TotalDevicePoolSize to inform the user of the
problem details. If this error is detected, program execution is halted.

This find resulted in the correction of about 10 tests due to now
realized miscalculated device pool sizes.