Save before huge refactor
The changes in this commit are a result of me trying to overcome a
directory size issue, but alas, all is lost. Trying to have concurrent
backup to mulitple devices each with possibly different filesystems
while making the directory structure generally usable by humans is a
hard problem.

I'd rather have multi-device streaming than preserved directory
structure. So I am going back to the roots of the project.

The problem started when I noticed copying from a variable blocksize
filesystem (ZFS) resulted in the backup process hitting a race condition
when the device ran out of space when it should have had more space
available.

On ZFS, the size of the problematic directories (with 772 files), is so:

3       /mnt/data/files/games/emulation/Atari 2600
774     /mnt/data/files/games/emulation/Atari 2600/ROMS

But, when these directories are copied to an EXT4 formatted device, they
become:

4096    /mnt/data/files/games/emulation/Atari 2600
36864   /mnt/data/files/games/emulation/Atari 2600/ROMS

EXT4 has a default blocksize of 4096 bytes.

Trying to calculate the directory size for a list of files contained in
a directory is possible, but would have made it impracticle for the
multitudes of filesystems currently available. Maintaining directory
size calculations for every possible filesytem a device can use is not
something I want to mantain.

So my next question was, well how does tar do it? Turns out, the
directory files have no size in a tar file. Which leads me to how this
project originally began: I simply copied the data in the files to the
destination device and gave each file a UUID name. With this method,
data is deterministic, and everything becomes much easier to handle. So
I will be refactoring back to the old way.