Fix sync file progress output
Slowed down the various for loops with time.Second / n pauses and checks
in the tracker.report() method. This allows the Go scheduler to give
other goroutines more resources to play with. It feels like things are
working much better now, and on top of all that, the gui is showing the
beginnings of actual file transfer progress (it's a little more
accurate).