Preliminary fix for out-of-space errors
When splitting a file on low device space, an extra two tenths of one
percent of the total device space will be subtracted from the split
file. This is to help ensure no out-of-pace errors occurr due to a
combination of number of files on the device and the information the
directory structure must contain to support the files.