UUID names for destination files
Simplifies everything. The destination file is saved using a UUID name.
Now it is not necessary to be concerned with preserving directory
structure. All this information is copied to each device in the catalog
output for safe keeping.

Data recovery will be next to impossible without this catalog file. It
is needed to recreate the directory structure.