Integrate termui source completely into conui
The customizations that are needed to termui to build the GDS GUI are
too great. It is much simpler to just import the source directly and
modify it to fit the needs of gds.

termui has an excellent layout engine for text based GUIs. Having to
redo all of this from scratch just so I can control panels would not be
practical.