Improve scrolling yet again
Make a scroll ring out of the device panels. For example, when there are
three devices, and enough room in the terminal to show two devices, and
the third device is selected, then the first device will be rendered
after the currently selected third device. This make it so the user can
see that there are other devices available if there is sufficient space
available to render atleast two device panels.