core/sync.go: Fix a very subtle bug with progress reporting
After a receive is made from the s.io.sizeWritn channel, s.io.sizeWritnTotal was
compared to the the size of the file "s.File.TotalSize". Since this comparison
was running in a separate goroutine, it would sometimes equal true (forcing an
exit of the loop) before the next receieve of the number of copied bytes could
be made from s.io.sizeWritn. This had the undesired affect of some bytes not
being reported as copied, throwing off the GUI and making it appear like a short
write had occurred.