context.go: Add PaddingPercentage
Not all devices can be filled to maximum capacity. The PaddingPercentage
configuration value sets the amount of bytes that will be left on the
device once the sync is complete. The default padding percentage is 1%
of defined space. Otherwise, the padding percentage is pulled from the
configuration file. a toplevel value set the padding for all devices. If
the value is specified for a device, that will overide the default and
toplevel value.

Implementing PaddingPercentage required a major change to how the
context is initialized.