GINORMOUS refactor: Catalog has been changed to FileIndex
The reason for the huge refactor: I wanted to simply check the sha1sum
of all the parts of a split file in a single hash. The catalog had the
the file parts, and the FileList had the source files. I would have had
to use both lists to check the hash, which would have made for ugly code.

So instead, the Catalog and FileList are merged, with each file
containing a list of destination files. This allows for testing the
sha1sum of split files easily. The cataloging process is much more
approachable now.

I also tackled the refactoring of synctest that has been on the back
burner for a while. It's so much better now and much more easier to wrap
my head around after being away from it for a while.