hashing_progress.go: Fix byte write reporting bug
The problem was fromm using pointers instead of values for HashFile. The hash goroutine passes a
pointer back to the calcFileIndexHashes() goroutine indicating a write has occurred. A check for
"hf.SizeWritn == hf.SizeTotal" is done, if true then "calcFileIndexHashes: RECEIVED: FILE WRITE
COMPLETE" is written to the log. In the time it took to receive the pointer, the io.Copy goroutine
has already finished writing the file, and on the completion of full write the pointer was sent
again, producing the second "hf.SizeWritn == hf.SizeTotal" causing the second occurrence of "FILE
WRITE COMPLETE" of output.

This meant the BytesWritnLast recorded to the location of the pointer was getting calculated twice.
In some cases it was more than the size of the file, in others, it was less. This led to incorrect
stat data in the hashing progress overlay.

A special test program was created in test/debug/reporting.go to help track down the problem using
the JSON formatter of the log output.